flow/wallclock: fix overflow on 32bit machines.
ts.tv_sec * 1000 will overflow on 32 bits, make sure that is 64 bits
before doing the multiply.

This was spotted by -fsanitize=undefined on my 32bits VM.

+++ src/modules/flow/wallclock/wallclock.c
@@ -212,7 +212,7 @@ wallclock_schedule_next(struct sol_flow_node *node)
 
     if (mdata->type == TIMEOUT_SECOND) {
         CLOCK_GETTIME_DO(timeout,
-            1000 - ((ts.tv_sec * 1000)
+            1000 - (((uint64_t)ts.tv_sec * 1000)
             + (ts.tv_nsec / 1000000)) % 1000);
         if (time_fail)
             goto err;
